Added placeholder for industrial_robot_status_interface
